+++ src/main.js
@@ -1,14 +1,14 @@
 import FiltersView from './view/filters-view.js';
 import TripInfoView from './view/trip-info-view.js';
 import {RenderPosition, render} from './render.js';
-import BoardPresenter from './presenter/board-presenter.js';
+import EventsPresenter from './presenter/events-presenter.js';
 
 const tripInfoContainer = document.querySelector('.trip-main');
 const filtersContainer = document.querySelector('.trip-controls__filters');
 const eventsContainer = document.querySelector('.trip-events');
-const boardPresenter = new BoardPresenter({boardContainer: eventsContainer});
+const eventsPresenter = new EventsPresenter({eventsContainer: eventsContainer});
 
 render(new TripInfoView(), tripInfoContainer, RenderPosition.AFTERBEGIN);
 render(new FiltersView(), filtersContainer);
 
-boardPresenter.init();
+eventsPresenter.init();
